requirements.txt
Django==4.2.7
Pillow==10.0.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
python-dotenv==1.0.0
django-allauth==0.57.0
mysqlclient==2.2.7 